Updated config options, finally got steamcmd working

+++ scripts/Server-Tools/Update-Tools.ps1
@@ -18,7 +18,7 @@ function AddUpdateLock
 
 function RemoveUpdateLock
 {
-  Remove-Item $lockFilePath
+  if (Test-Path $lockFilePath) { Remove-Item $lockFilePath }
 }
 
 function InstalledVersion
